Makefile: Support for alternate tgtd executable name
When we are building an osd-target executable, we should
not conflict with original tgtd name, which might be on
the system by means of a distribution package.

Also the osd-target Makefile has options to build three
different backends, As a compile time switch. It is
dangerous to mix-up between these because some backend
files are the same and some not. Better call each executable
different. So far we use three options:

otgtd		- The default original backend (highly recommended)
pan_tgtd	- For use on Panasas OSDFS mount, usually compiled for
		  FreeBSD
pansim_tgtd	- For Panasas simulator OSD targets. Can be a Linux or
		  FreeBSD build.

.gitignore is updated for these alternate names.

From here on we'll use these new names in script files. Everyone should
update their scripts to use otgtd when running an osd-target.

Signed-off-by: Boaz Harrosh <bharrosh@panasas.com>